requirements.txt
asgiref==3.8.1
boto3==1.34.100
channels==4.1.0
channels-redis==4.2.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
Pillow
python-decouple==3.8
social-auth-app-django==5.4.0
whitenoise==6.6.0

